load react component lazily in so management section (#64285)
* load react component lazily in so management section

* remove unused imports
